Move SDK init from ViewController to AppDelegate (also moves the TapResearchSDKDelegate function to AppDelegate)
