+++ ios/ObjC+UIKit/IntegrationDemo_ObjC_UIKit/AppDelegate.h
@@ -6,7 +6,8 @@
 //
 
 #import <UIKit/UIKit.h>
+#import <TapResearchSDK/TapResearchSDK.h>
 
-@interface AppDelegate : UIResponder <UIApplicationDelegate>
+@interface AppDelegate : UIResponder <UIApplicationDelegate, TapResearchSDKDelegate>
 
 @end
